BE::Time::Timer's elapsed() returns uint64_t. No need to store it in BE::Framework::API as a double.
[git-p4: depot-paths = "//depot/projects/EVALUATION/Main/common/": change = 8276]

+++ src/include/be_framework_api.h
@@ -60,7 +60,7 @@ namespace BiometricEvaluation
 				Result();
 
 				/** Time elapsed while calling operation. */
-				double elapsed;
+				uint64_t elapsed;
 				/**
 				 * @brief
 				 * Value returned from operation.
